requirements.txt
python-wordpress-xmlrpc==2.3
soundcloud==0.5.0
